Split epee/string_tools.h and encapsulate boost::lexical_cast
