+++ contrib/epee/include/misc_os_dependent.h
@@ -47,7 +47,7 @@
 #endif
 
 #include <iostream>
-#include <boost/lexical_cast.hpp>
+#include <ctime>
 
 #pragma once 
 namespace epee
@@ -115,15 +115,7 @@ namespace misc_utils
 
 	}
 
-
-	inline std::string get_thread_string_id()
-	{
-#if defined(_WIN32)
-		return boost::lexical_cast<std::string>(GetCurrentThreadId());
-#elif defined(__GNUC__)  
-		return boost::lexical_cast<std::string>(pthread_self());
-#endif
-	}
+	std::string get_thread_string_id();
 
 	inline bool get_gmt_time(time_t t, struct tm &tm)
 	{
