+++ contrib/epee/include/reg_exp_definer.h
@@ -29,6 +29,7 @@
 #define _REG_EXP_DEFINER_H_
 
 #include <boost/interprocess/detail/atomic.hpp>
+#include <boost/regex.hpp>
 #include "syncobj.h"
 
 namespace epee
