+++ contrib/epee/include/string_tools_lexical.h
@@ -24,38 +24,68 @@
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 // 
 
+#ifndef _STRING_TOOLS_LEXICAL_H_
+#define _STRING_TOOLS_LEXICAL_H_
 
-#ifndef _TINY_INI_H_
-#define _TINY_INI_H_
+#include "warnings.h"
+#include "storages/parserse_base_utils.h"
+#include <boost/lexical_cast.hpp> // A heavy header, that was extracted from the rest
 
-#include "string_tools.h"
+#ifndef OUT
+	#define OUT
+#endif
 
 namespace epee
 {
-namespace tiny_ini
+namespace string_tools
 {
+PUSH_WARNINGS
+DISABLE_GCC_WARNING(maybe-uninitialized)
+    template<class XType>
+    inline bool get_xtype_from_string(OUT XType& val, const std::string& str_id)
+    {
+        if (std::is_integral<XType>::value && !std::numeric_limits<XType>::is_signed && !std::is_same<XType, bool>::value)
+        {
+            for (char c : str_id)
+            {
+                if (!epee::misc_utils::parse::isdigit(c))
+                    return false;
+            }
+        }
 
-	bool get_param_value(const std::string& param_name, const std::string& ini_entry, std::string& res);
-	inline std::string get_param_value(const std::string& param_name, const std::string& ini_entry)
-	{
-		std::string buff;
-		get_param_value(param_name, ini_entry, buff);
-		return buff;
-	}
+        try
+        {
+            val = boost::lexical_cast<XType>(str_id);
+            return true;
+        }
+        catch(const std::exception& /*e*/)
+        {
+            //const char* pmsg = e.what();
+            return false;
+        }
+        catch(...)
+        {
+            return false;
+        }
 
-	template<class T>
-		bool get_param_value_as_t(const std::string& param_name, const std::string& ini_entry, T& res)
-	{
-		std::string str_res = get_param_value(param_name, ini_entry);
-	
-		string_tools::trim(str_res);
-		if(!str_res.size())
-			return false;
-		
-		return string_tools::get_xtype_from_string(res, str_res);
-	}
+        return true;
+    }
+POP_WARNINGS
 
+    template<class XType>
+    inline bool xtype_to_string(const  XType& val, std::string& str)
+    {
+        try
+        {
+            str = boost::lexical_cast<std::string>(val);
+        }
+        catch(...)
+        {
+            return false;
+        }
+
+        return true;
+    }
 }
 }
-
-#endif //_TINY_INI_H_
+#endif //_STRING_TOOLS_LEXICAL_H_
