+++ contrib/epee/src/http_base.cpp
@@ -24,38 +24,48 @@
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 // 
 
-
-#ifndef _TINY_INI_H_
-#define _TINY_INI_H_
-
+#include "net/http_base.h"
+#include "memwipe.h"
 #include "string_tools.h"
+
+#include <boost/regex.hpp>
+#include <string>
+#include <utility>
+
+#undef MONERO_DEFAULT_LOG_CATEGORY
+#define MONERO_DEFAULT_LOG_CATEGORY "net.http"
 
 namespace epee
 {
-namespace tiny_ini
+namespace net_utils
 {
+namespace http
+{
+    std::string get_value_from_fields_list(const std::string& param_name, const net_utils::http::fields_list& fields)
+    {
+        fields_list::const_iterator it = fields.begin();
+        for(; it != fields.end(); it++)
+            if(!string_tools::compare_no_case(param_name, it->first))
+                break;
 
-	bool get_param_value(const std::string& param_name, const std::string& ini_entry, std::string& res);
-	inline std::string get_param_value(const std::string& param_name, const std::string& ini_entry)
-	{
-		std::string buff;
-		get_param_value(param_name, ini_entry, buff);
-		return buff;
-	}
+        if(it==fields.end())
+            return std::string();
 
-	template<class T>
-		bool get_param_value_as_t(const std::string& param_name, const std::string& ini_entry, T& res)
-	{
-		std::string str_res = get_param_value(param_name, ini_entry);
-	
-		string_tools::trim(str_res);
-		if(!str_res.size())
-			return false;
-		
-		return string_tools::get_xtype_from_string(res, str_res);
-	}
+        return it->second;
+    }
 
+    std::string get_value_from_uri_line(const std::string& param_name, const std::string& uri)
+    {
+        std::string buff = "([\\?|&])";
+        buff += param_name + "=([^&]*)";
+        boost::regex match_param(buff.c_str(), boost::regex::icase | boost::regex::normal);
+        boost::smatch	result;
+        if(boost::regex_search(uri, result, match_param, boost::match_default) && result[0].matched) 
+        {
+            return result[2];
+        }
+        return std::string();
+    }
 }
 }
-
-#endif //_TINY_INI_H_
+}
