+++ src/daemon/command_parser_executor.cpp
@@ -31,6 +31,7 @@
 #include "net/parse.h"
 #include "daemon/command_parser_executor.h"
 #include <boost/filesystem.hpp>
+#include <boost/algorithm/string/predicate.hpp>
 
 #undef MONERO_DEFAULT_LOG_CATEGORY
 #define MONERO_DEFAULT_LOG_CATEGORY "daemon"
