+++ src/net/i2p_address.cpp
@@ -38,7 +38,7 @@
 #include "net/error.h"
 #include "serialization/keyvalue_serialization.h"
 #include "storages/portable_storage.h"
-#include "string_tools.h"
+#include "string_tools_lexical.h"
 
 namespace net
 {
