+++ tests/unit_tests/get_xtype_from_string.cpp
@@ -31,6 +31,7 @@
 #include "gtest/gtest.h"
 
 #include <string_tools.h>
+#include <string_tools_lexical.h>
 
 using namespace epee::string_tools;
